Refactor SinglePost to Post
